apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
google-cloud-storage==2.16.0
google-cloud-firestore==2.16.0
reportlab==4.2.0
python-docx==1.1.2
